% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discover_datasets.R
\name{discover_datasets}
\alias{discover_datasets}
\title{Discover Available Datasets and Variables}
\usage{
discover_datasets(source = "cbn", dataset = NULL)
}
\arguments{
\item{source}{Character. Data source (cbn, nbs, wb, ngx).}

\item{dataset}{Character. Dataset identifier.}
}
\description{
Discover Available Datasets and Variables
}
