% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naijr_cache_clear.R
\name{naijr_cache_clear}
\alias{naijr_cache_clear}
\title{Clear opennaijR cache}
\usage{
naijr_cache_clear(dataset = NULL, all = FALSE)
}
\arguments{
\item{dataset}{Character scalar. Dataset name (e.g., "fx_reserves").}

\item{all}{Logical. If TRUE, clears the entire cache.}
}
\value{
Invisibly TRUE if successful, FALSE otherwise.
}
\description{
Clear opennaijR cache
}
\examples{
\dontrun{
naijr_cache_clear("fx_reserves")
naijr_cache_clear(all = TRUE)
}
}
