% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naijr_cache_info.R
\name{naijr_cache_info}
\alias{naijr_cache_info}
\title{Inspect metadata for a cached dataset}
\usage{
naijr_cache_info(name)
}
\arguments{
\item{name}{Character scalar. Name of the cached dataset (as returned by
\code{naijr_cache_list()})}
}
\value{
A list containing metadata for the dataset.
}
\description{
Retrieves metadata (e.g., creation time, size) for a cached dataset.
}
\examples{
\dontrun{
pins <- naijr_cache_list()
naijr_cache_info(pins[1])
}
}
