% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naijr_cache_list.R
\name{naijr_cache_list}
\alias{naijr_cache_list}
\title{List cached opennaijR datasets}
\usage{
naijr_cache_list()
}
\value{
Character vector of cached pin names.
}
\description{
Lists all datasets currently cached by opennaijR.
}
\examples{
\dontrun{
naijr_cache_list()
}
}
